requirements.txt
fastapi
uvicorn
pydantic
pyyaml
streamlit
requests
